+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Added
 - Add ability to use relative symlinks
+- Add ability to pass allowed servers via flag (eg. `deploy --on="production" --server="1"`)
 
 ### Fixed
 - Fixed symlink issues on non-GNU/Linux OSes